requirements.txt
google-api-python-client==2.65.0
pyttsx3==2.90
pytz==2022.5
selenium==4.5.0
SpeechRecognition==3.8.1
xlwt==1.3.0
xlrd==2.0.1
yagmail==0.15.293
